clean up duplicate code
